Test getClient
