+++ .env.test
@@ -1,23 +1,16 @@
-# direnv will make a copy of this file as an .env file
-# Follow the instructions in README.md as how to add your personal
-# tokens and secrets in this file
-
-# GitHub App ID
+# GitHub
+FORCE_USER_TOKEN_GITHUB_CLIENT="false"
+GH_USER_TOKEN="ghp_BLAHBLAHBLAH"
 GH_APP_IDENTIFIER="1234"
-# Github App Webhook Secret to verify requests come from GitHub
 GH_WEBHOOK_SECRET="webhooksecret"
-# GitHub App Secret Key
-# NOTE: this *must* be on a single line, otherwise it will break
-# reading the secret key.
-GH_APP_SECRET_KEY='-----BEGIN RSA PRIVATE KEY----------END RSA PRIVATE KEY-----'
+GH_APP_SECRET_KEY="top \nsecret\n key"
 
-# Slack signing secret to verify requests come from Slack
+# Slack
 SLACK_SIGNING_SECRET="slacksigningsecret"
-# Slack Bot user access token -> "Bot User OAuth Token"
 SLACK_BOT_USER_ACCESS_TOKEN="slackbotaccesstoken"
-# The Slack bot's App ID
 SLACK_BOT_APP_ID="5678"
 
+# Other
 OWNER="getsentry"
 SENTRY_REPO="sentry"
 GETSENTRY_REPO="getsentry"
@@ -28,7 +21,6 @@ FEED_DEPLOY_CHANNEL_ID="9101112"
 FEED_DEV_INFRA_CHANNEL_ID="13141516"
 FEED_ENGINEERING_CHANNEL_ID="33333333"
 DISABLE_GITHUB_METRICS="false"
-FORCE_USER_TOKEN_GITHUB_CLIENT=
 FORCE_GET_USER_BY_SLACK_ID=
 TEAM_OSPO_CHANNEL_ID=
 SUPPORT_CHANNEL_ID=
