--- src/api/github/__mocks__/getClient.ts
@@ -1,181 +1,0 @@
-import MockCompareCommits from '@test/compareCommits.json';
-import { workflow_run_job } from '@test/payloads/github/workflow_run_job';
-
-import { ClientType } from '../clientType';
-
-import { MockOctokitError } from './mockError';
-
-function mockClient() {
-  return {
-    paginate: jest.fn(),
-    actions: {
-      listWorkflowRunsForRepo: jest.fn(),
-      cancelWorkflowRun: jest.fn(),
-      getJobForWorkflowRun: jest.fn(async ({ job_id }) => {
-        return { data: workflow_run_job({ job_id }) };
-      }),
-    },
-    checks: {
-      listForRef: jest.fn(),
-      listAnnotations: jest.fn(async () => {
-        return {
-          data: [
-            {
-              path: 'tests/snuba/rules/conditions/test_event_frequency.py',
-              blob_href:
-                'https://github.com/getsentry/sentry/blob/83ef9b927cbb822febbdf75e5e05dd40afb187cf/tests/snuba/rules/conditions/test_event_frequency.py',
-              start_line: 570,
-              start_column: null,
-              end_line: 570,
-              end_column: null,
-              annotation_level: 'failure',
-              title:
-                'tests/snuba/rules/conditions/test_event_frequency.py#L570',
-              message:
-                'EventFrequencyPercentConditionTestCase.test_one_hour_with_events\n' +
-                '\n' +
-                'AssertionError',
-              raw_details: null,
-            },
-            {
-              path: '.github',
-              blob_href:
-                'https://github.com/getsentry/sentry/blob/83ef9b927cbb822febbdf75e5e05dd40afb187cf/.github',
-              start_line: 1,
-              start_column: null,
-              end_line: 1,
-              end_column: null,
-              annotation_level: 'failure',
-              title: '.github#L1',
-              message: 'Process completed with exit code 2.',
-              raw_details: null,
-            },
-          ],
-        };
-      }),
-    },
-    git: {
-      getCommit: jest.fn(),
-    },
-    teams: {
-      getByName: jest.fn(async function (payload) {
-        if (
-          payload.team_slug === 'product-owners-test' ||
-          payload.team_slug === 'product-owners-rerouted'
-        ) {
-          return { status: 200, data: {} };
-        }
-        throw new MockOctokitError(404);
-      }),
-    },
-    graphql: jest.fn(),
-    issues: {
-      _labels: new Set([]),
-      _comments: [],
-      addLabels: jest.fn(async function (payload) {
-        for (const name of payload.labels) {
-          this._labels.add(name);
-        }
-      }),
-      removeLabel: jest.fn(async function (payload) {
-        this._labels.delete(payload.name);
-      }),
-      createComment: jest.fn(async function (payload) {
-        this._comments.push(payload.body);
-      }),
-      update: jest.fn(),
-    },
-    orgs: {
-      checkMembershipForUser: jest.fn(async function (x) {
-        let status = 302;
-        if (x.org === 'Enterprise' || x.org === null) {
-          if (x.username === 'Picard' || x.username === 'Troi') {
-            return { status: 204 };
-          } else {
-            status = 404;
-          }
-        }
-        throw new MockOctokitError(status);
-      }),
-    },
-    pulls: {
-      get: jest.fn(),
-    },
-    repos: {
-      getCommit: jest.fn(),
-
-      listPullRequestsAssociatedWithCommit: jest.fn(),
-
-      getContent: jest.fn(),
-
-      compareCommits: jest.fn(({ base, head }) => {
-        // If base is older than head, then status will be ahead
-        // For tests it might be easier to think of base/head as incrementing
-        // ints (representing age) instead of a hash
-        const isAhead = !base && !head ? true : base < head ? true : false;
-        const commits = MockCompareCommits.data.commits;
-        const oldestCommit = {
-          ...commits[0],
-          sha: isAhead ? base ?? commits[0].sha : null,
-        };
-        const newestCommit = {
-          ...commits[1],
-          sha: isAhead ? head ?? commits[1].sha : null,
-        };
-        const mockCommits = isAhead ? [oldestCommit, newestCommit] : [];
-        const numCommits = mockCommits.length;
-
-        return {
-          ...MockCompareCommits,
-          status: 200,
-          data: {
-            // If behind, there will be no commits in response
-            commits: mockCommits,
-            status: isAhead ? 'ahead' : base === head ? 'identical' : 'behind',
-            ahead_by: numCommits,
-            behindBy: isAhead ? 0 : 2,
-            total_commits: numCommits,
-          },
-        };
-      }),
-    },
-
-    request: jest.fn(async (URL, x) => {
-      if (URL.includes && URL.includes('/GTM/')) {
-        let status = 302;
-        if (x.org === 'Enterprise' || x.org === null) {
-          if (x.username === 'Troi') {
-            return { status: 200 };
-          } else {
-            status = 404;
-          }
-        }
-        throw new MockOctokitError(status);
-      }
-      return {};
-    }),
-  };
-}
-
-let userMock;
-const appMocks = {};
-
-export async function getClient(type: ClientType, org?: string) {
-  if (type == ClientType.User) {
-    if (userMock) {
-      return userMock;
-    }
-    userMock = mockClient();
-    return userMock;
-  } else if (type == ClientType.App) {
-    if (!org) {
-      throw Error('Org required for mock getClient()');
-    }
-    if (appMocks[org]) {
-      return appMocks[org];
-    }
-
-    appMocks[org] = mockClient();
-    return appMocks[org];
-  }
-}
