+++ src/api/github/octokitWithRetries.ts
@@ -1,0 +1,3 @@
+import { retry } from '@octokit/plugin-retry';
+import { Octokit } from '@octokit/rest';
+export const OctokitWithRetries = Octokit.plugin(retry);
